+++ src/molnetpack/model.py
@@ -105,16 +105,132 @@ class MolConv(nn.Module):
 	def __repr__(self):
 		return self.__class__.__name__ + ' k = ' + str(self.k) + ' (' + str(self.in_dim) + ' -> ' + str(self.out_dim) + ')'
 
+class MolConv2(nn.Module):
+	def __init__(self, in_dim, out_dim, k, remove_xyz=False):
+		super(MolConv2, self).__init__()
+		self.k = k
+		self.in_dim = in_dim
+		self.out_dim = out_dim
+		self.remove_xyz = remove_xyz
+
+		self.dist_ff = nn.Sequential(nn.Conv2d(1, 1, kernel_size=1, bias=False),
+								nn.BatchNorm2d(1),
+								nn.Sigmoid())
+
+		if remove_xyz: 
+			self.center_ff = nn.Sequential(nn.Conv2d(in_dim-3, in_dim+k-3, kernel_size=1, bias=False),
+								nn.BatchNorm2d(in_dim+k-3),
+								nn.Sigmoid())
+			self.update_ff = nn.Sequential(nn.Conv2d(in_dim+k-3, out_dim, kernel_size=1, bias=False),
+								nn.BatchNorm2d(out_dim),
+								nn.LeakyReLU(negative_slope=0.02))
+		else:
+			self.center_ff = nn.Sequential(nn.Conv2d(in_dim, in_dim+k, kernel_size=1, bias=False),
+								nn.BatchNorm2d(in_dim+k),
+								nn.Sigmoid())
+			self.update_ff = nn.Sequential(nn.Conv2d(in_dim+k, out_dim, kernel_size=1, bias=False),
+								nn.BatchNorm2d(out_dim),
+								nn.LeakyReLU(negative_slope=0.02))
+
+		self._reset_parameters()
+
+	def _reset_parameters(self): 
+		for m in self.modules(): 
+			if isinstance(m, (nn.Conv1d, nn.Conv2d, nn.Linear)):
+				nn.init.kaiming_normal_(m.weight, a=0.2, mode='fan_in', nonlinearity='leaky_relu')
+			
+			elif isinstance(m, (nn.BatchNorm1d, nn.BatchNorm2d, nn.GroupNorm)): 
+				nn.init.constant_(m.weight, 1)
+				nn.init.constant_(m.bias, 0)
+
+	def forward(self, x: torch.Tensor, 
+						idx_base: torch.Tensor) -> torch.Tensor: 
+		dist, gm2, feat_c, feat_n = self._generate_feat(x, idx_base, k=self.k, remove_xyz=self.remove_xyz) 
+		'''Returned features: 
+		dist: torch.Size([batch_size, 1, point_num, k])
+		gm2: torch.Size([batch_size, k, point_num, k]) 
+		feat_c: torch.Size([batch_size, in_dim, point_num, k]) 
+		feat_n: torch.Size([batch_size, in_dim, point_num, k])
+		'''
+		feat_n = torch.cat((feat_n, gm2), dim=1) # torch.Size([batch_size, in_dim+k, point_num, k])
+		feat_c = self.center_ff(feat_c)
+
+		w = self.dist_ff(dist)
+
+		feat = w * feat_n + feat_c
+		feat = self.update_ff(feat)
+		feat = feat.mean(dim=-1, keepdim=False)
+		return feat
+
+	def _generate_feat(self, x: torch.Tensor, 
+								idx_base: torch.Tensor, 
+								k: int, 
+								remove_xyz: bool) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]: 
+		batch_size, num_dims, num_points = x.size()
+		
+		# local graph (knn)
+		inner = -2*torch.matmul(x.transpose(2, 1), x)
+		xx = torch.sum(x**2, dim=1, keepdim=True)
+		pairwise_distance = -xx - inner - xx.transpose(2, 1)
+		dist, idx = pairwise_distance.topk(k=k, dim=2) # (batch_size, num_points, k)
+		dist = - dist
+
+		idx = idx + idx_base
+		idx = idx.view(-1)
+
+		x = x.transpose(2, 1).contiguous() # (batch_size, num_points, num_dims) -> (batch_size*num_points, num_dims) 
+		# print('_double_gram_matrix (x):', torch.any(torch.isnan(x)))
+		graph_feat = x.view(batch_size*num_points, -1)[idx, :]
+		# print('_double_gram_matrix (graph_feat):', torch.any(torch.isnan(graph_feat)))
+		graph_feat = graph_feat.view(batch_size, num_points, k, num_dims)
+
+		# gram matrix
+		gm_matrix = torch.matmul(graph_feat, graph_feat.permute(0, 1, 3, 2))
+		# print('_double_gram_matrix (gm_matrix):', torch.any(torch.isnan(gm_matrix)))
+		# gm_matrix = F.normalize(gm_matrix, dim=1) 
+
+		# double gram matrix
+		sub_feat = gm_matrix[:, :, :, 0].unsqueeze(3)
+		sub_gm_matrix = torch.matmul(sub_feat, sub_feat.permute(0, 1, 3, 2))
+		sub_gm_matrix = F.normalize(sub_gm_matrix, dim=1) 
+		# print('_double_gram_matrix (sub_gm_matrix):', torch.any(torch.isnan(sub_gm_matrix)))
+		
+		x = x.view(batch_size, num_points, 1, num_dims).repeat(1, 1, k, 1)
+
+		if remove_xyz:
+			return dist.unsqueeze(3).permute(0, 3, 1, 2).contiguous(), \
+					sub_gm_matrix.permute(0, 3, 1, 2).contiguous(), \
+					x[:, :, :, 3:].permute(0, 3, 1, 2).contiguous(), \
+					graph_feat[:, :, :, 3:].permute(0, 3, 1, 2).contiguous()
+		else:
+			return dist.unsqueeze(3).permute(0, 3, 1, 2).contiguous(), \
+					sub_gm_matrix.permute(0, 3, 1, 2).contiguous(), \
+					x.permute(0, 3, 1, 2).contiguous(), \
+					graph_feat.permute(0, 3, 1, 2).contiguous()
+	
+	def __repr__(self):
+		return self.__class__.__name__ + ' k = ' + str(self.k) + ' (' + str(self.in_dim) + ' -> ' + str(self.out_dim) + ')'
+
+
+
 class Encoder(nn.Module):
 	def __init__(self, in_dim, layers, emb_dim, k): 
 		super(Encoder, self).__init__()
 		self.emb_dim = emb_dim
-		self.hidden_layers = nn.ModuleList([MolConv(in_dim=in_dim, out_dim=layers[0], k=k, remove_xyz=True)])
+		# v1: 
+		# self.hidden_layers = nn.ModuleList([MolConv(in_dim=in_dim, out_dim=layers[0], k=k, remove_xyz=True)])
+		# for i in range(1, len(layers)): 
+		# 	if i == 1:
+		# 		self.hidden_layers.append(MolConv(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
+		# 	else:
+		# 		self.hidden_layers.append(MolConv(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
+		# v2: 
+		self.hidden_layers = nn.ModuleList([MolConv2(in_dim=in_dim, out_dim=layers[0], k=k, remove_xyz=True)])
 		for i in range(1, len(layers)): 
 			if i == 1:
-				self.hidden_layers.append(MolConv(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
+				self.hidden_layers.append(MolConv2(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
 			else:
-				self.hidden_layers.append(MolConv(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
+				self.hidden_layers.append(MolConv2(in_dim=layers[i-1], out_dim=layers[i], k=k, remove_xyz=False))
 		
 		self.conv = nn.Sequential(nn.Conv1d(emb_dim, emb_dim, kernel_size=1, bias=False),
 								   nn.BatchNorm1d(emb_dim),